Re-split classes
