+++ classes/reversestring.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+def reverse(str)
+  str.reverse
+end
